fix: Node names have to be unique
"Kind" was not part of node name before. Why? Well it wouldn't look
right in the graph output if the nodes were called "Deployment_xxxx"
instead of "xxxx". But we are explicityl setting the "label" attr now
which control what is seen in the output so adding kind doesn't matter.